Initial Commit: Bitbucket plugin for adding groups to pull-request
See README.md for explanation & screenshots

Initial commit for a Bitbucket plugin which adds a button to the Pull Request Creation Page.
Clicking this button will open a dialog with a list of groups that have permission to the repository or project.

The user can check a box next to each group (or click select all). Members from each group will be added as a reviewer.

Tested on Bitbucket 7.1.1 & 6.8.2