PullRequestGroups: Global Config Page, Work with Forks
* Add a global configuration page for PullRequestGroups with two
options: Set the maximum group size value, ability to hide truncated
groups.

* Documentation for API endpoints

* Use i18nService for strings

* Work with forks: Before, the script would get the current project,
current repository. This was fine if the PR source & target branch were
in the same repository. But if source was a different repository, then
it would fetch the groups of the source (current repository) - not the target.
So now the script will fetch the groups of the selected PR target

* Support for Bitbucket 6 to latest

* Use SpringScanner Verison 2

* Release version 0.1.5